app/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile app/requirements.in
blinker==1.9.0
    # via flask
click==8.1.8
    # via flask
colorama==0.4.6
    # via click
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
flask==3.1.0
    # via -r app/requirements.in
fonttools==4.57.0
    # via matplotlib
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
kiwisolver==1.4.8
    # via matplotlib
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
matplotlib==3.10.3
    # via -r app/requirements.in
numpy==2.2.5
    # via
    #   contourpy
    #   matplotlib
    #   pandas
packaging==25.0
    # via matplotlib
pandas==2.2.3
    # via -r app/requirements.in
pillow==11.2.1
    # via matplotlib
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.1.0
    # via -r app/requirements.in
pytz==2025.2
    # via pandas
redis==6.0.0
    # via -r app/requirements.in
six==1.17.0
    # via python-dateutil
tzdata==2025.2
    # via pandas
werkzeug==3.1.3
    # via flask
